+++ back/prisma/seed.ts
@@ -1,0 +1,12 @@
+import fs from "fs";
+import path from "path";
+
+(() => {
+  try {
+    fs.accessSync(path.join(__dirname, "seed.dev.ts"));
+  } catch (err) {
+    return;
+  }
+
+  import("./seed.dev");
+})();