+++ back/src/companies/resolvers/queries/companyInfos.ts
@@ -32,12 +32,13 @@ export async function getCompanyInfos(siret: string): Promise<CompanyPublic> {
       contactEmail
       contactPhone
       website
+      ecoOrganismeAgreements
     }
   `;
 
   type CompanyFragment = Pick<
     Company,
-    "contactEmail" | "contactPhone" | "website"
+    "contactEmail" | "contactPhone" | "website" | "ecoOrganismeAgreements"
   >;
 
   // retrieves trackdechets public CompanyInfo