+++ back/src/scalars/index.ts
@@ -1,0 +1,5 @@
+import { URLResolver } from "graphql-scalars";
+
+export default {
+  URL: URLResolver
+};