+++ front/src/account/accountCompanyAdd/AccountCompanyAddEcoOrganisme.tsx
@@ -1,0 +1,60 @@
+import React from "react";
+import { FaTrash } from "react-icons/fa";
+import { Field, FieldArray, useField } from "formik";
+import RedErrorMessage from "common/components/RedErrorMessage";
+import styles from "../AccountCompanyAdd.module.scss";
+
+export default function AccountCompanyAddEcoOrganisme() {
+  const fieldProps = { name: "ecoOrganismeAgreements" };
+  const [field] = useField<string[]>(fieldProps);
+
+  return (
+    <div className={styles.field}>
+      <label className={`text-right ${styles.bold}`}>
+        Agréments éco-organisme
+      </label>
+      <div className={styles.field__value}>
+        <table>
+          <tbody>
+            <tr>
+              <td>URL</td>
+              <td>
+                <FieldArray {...fieldProps}>
+                  {({ push, remove }) => (
+                    <>
+                      {field.value.map((url, index) => (
+                        <div key={index} className={styles.inputGroup}>
+                          <Field
+                            type="url"
+                            name={`ecoOrganismeAgreements.${index}`}
+                            className={styles.inputGroupInput}
+                          />
+                          <button
+                            type="button"
+                            className="button-outline small warning"
+                            onClick={() => remove(index)}
+                            aria-label="Supprimer"
+                          >
+                            <FaTrash />
+                          </button>
+                        </div>
+                      ))}
+                      <button
+                        type="button"
+                        className="button-outline primary"
+                        onClick={() => push("")}
+                      >
+                        Ajouter un agrément
+                      </button>
+                    </>
+                  )}
+                </FieldArray>
+                <RedErrorMessage name="ecoOrganismeAgreements" />
+              </td>
+            </tr>
+          </tbody>
+        </table>
+      </div>
+    </div>
+  );
+}